update repository views
* Backend: move to JAVA 17
* Frontend: change naming and update Home, Paper views
