+++ Backend/src/main/java/com/gist/cithub/backend/BackendApplication.java
@@ -7,7 +7,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @MapperScan({"com.gist.cithub.backend.Repo.dao","com.gist.cithub.backend.Tools.dao","com.gist.cithub.backend.User.dao"})
 @SpringBootApplication
 public class BackendApplication {
-
     public static void main(String[] args) {
         SpringApplication.run(BackendApplication.class, args);
     }
