+++ Frontend/src/ComponentCommon/Foot.vue
@@ -1,39 +1,17 @@
 
 <template>
-        <footer class="footer pt-3  ">
-        <div class="container-fluid">
-          <div class="row align-items-center justify-content-lg-between">
-            <div class="col-lg-6 mb-lg-0 mb-4">
-              <div class="copyright text-center text-sm text-muted text-lg-start">
-
-                made with <i class="fa fa-heart"></i> by
-                <a href="https://gist.nju.edu.cn/" class="font-weight-bold" target="_blank">Gist Group member Jiantu Xiang</a>
-                for CitHub .
-              </div>
-            </div>
-            <div class="col-lg-6">
-              <ul class="nav nav-footer justify-content-center justify-content-lg-end">
-                <li class="nav-item">
-                  <a href="https://gist.nju.edu.cn/" class="nav-link text-muted" target="_blank">Gist Group, Department of Computer Science and Technology, NanJing University</a>
-                </li>
-                <!-- <li class="nav-item">
-                  <a href="https://gist.nju.edu.cn/" class="nav-link text-muted" target="_blank">About Us</a>
-                </li> -->
-                <!-- <li class="nav-item">
-                  <a href="https://www.creative-tim.com/license" class="nav-link pe-0 text-muted" target="_blank">License</a>
-                </li> -->
-              </ul>
-            </div>
+  <footer class="footer pt-3">
+    <div class="container-fluid">
+      <div class="row align-items-center justify-content-lg-between">
+        <div class="col-lg-6 mb-lg-0 mb-4">
+          <div class="copyright text-center text-sm text-muted text-lg-start">
+            GIST @ Nanjing Univeristy
           </div>
         </div>
-      </footer>
+      </div>
+    </div>
+  </footer>
 </template>
 
-<script  setup>
-
-</script>
-
-
-<style scoped>
-
-</style>
+<script setup></script>
+<style scoped></style>
