+++ Frontend/src/ComponentCommon/Navbar.vue
@@ -5,7 +5,6 @@
     <div class="container-fluid py-1 px-3">
       <nav aria-label="breadcrumb">
         <ol class="breadcrumb bg-transparent mb-0 pb-0 pt-1 px-0 me-sm-6 me-5">
-
           <li class="breadcrumb-item text-sm" @click="jumpToCithub" @mouseover="isHovered = true"
             @mouseleave="isHovered = false" style="cursor:pointer;" id="liCithub">
             <span v-show="isHovered" class="opacity-5 text-white hovered">CitHub</span>
@@ -15,10 +14,12 @@
           <li class="breadcrumb-item text-sm text-white active" aria-current="page">{{ moduleStore.CurrentModule }}</li>
         </ol>
       </nav>
+
+
       <div class="collapse navbar-collapse mt-sm-0 mt-2 me-md-0 me-sm-4" id="navbar">
         <div class="ms-md-auto pe-md-3 d-flex align-items-center">
         </div>
-        <ul class="navbar-nav  justify-content-end">
+        <ul class="navbar-nav justify-content-end">
           <li class="nav-item d-flex align-items-center">
             <a href="javascript:;" class="nav-link text-white font-weight-bold px-0">
               <i class="fa fa-user me-sm-1"></i>
