+++ Frontend/src/repository/ChartsDisplay.vue
