+++ Frontend/src/repository/RepositoryRouter.vue
@@ -2,7 +2,7 @@
 <router-view></router-view>
 </template>
 
-<script  setup>
+<script setup>
 
 </script>
