+++ Frontend/src/repository/components/ArgonBadge.vue
